Task_4/requirements.txt
asgiref==3.9.1
Django==3.2.25
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.7.0
pillow==11.3.0
pymongo==3.12.3
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.2.4
typing_extensions==4.14.1
tzdata==2025.2
